use move only functions to pass in getter for paths
we're doing this to make sure the paths are fetched in the async thread instead of in the main thread.
